Use column families to store chain data
We want blocks, txs, inputs and outputs in separate column families so
rocks can compress them individually. We also don't store tx data in
the block values itself as it is possible txs can be repeated in
uncles, we instead store txs separately, and track blockhash -> txid
index.

We even split out inputs and outputs in a separate column family as
this way we can track spent/unspent outputs without having to load the
entire tx from rocks.

The column families options are default for now and depending on the
performance we can tweak those.