Refactor ShareBlock and computed blockhash
We were using the bitcoin block's blockhash to track the chain, which was a mistake. We need to compute a hash of the block including the header and the transactions of p2pool.

We introduce a new type ShareBlockHash which enables us to use the type system to differentiate between a bitcoin BlockHash and the pool's blockhash.

We also introduce a number of new chain and store APIs to get and store shares at height. We introduced an index from height to blockhash. Using this index we simplified the locator queries and are thus now able to efficiently respond to GetHeaders and GetBlocks.

Massive commit, but this was a big refactor and new features leaked in too. :pray: