Create new connection for each gbt
bitcoind's RPC has very low timeout limits and it doesn't send a
connection termination when a connection is killed.

For the sake of making things reliable as opposed to optimised, which
can occassionally break, we are taking the approach of creating a new
http connection for each getblocktemplate and blocksubmit request.