Use share diff from block hash and mainnet max
Difficulty tracking in pools is the precise work in the header hash and doesn't depend on nbits field. The mainnet max is used for all chains, and is divided by work derived from header hash to get the difficulty of the share. This will contrast the chain's difficulty, but we want to capture exactly how much work is is captured by a share independent of the chain's current difficulty, specified in the block template nbits field.