Handle unaccounted for shares only once
When updated dsps for different windows we were
losing track of unaccounted for shares. We now do
that once and remember the total difficulty to use
for all windows.

We also had a small error in elapsed time. We were
using min instead of max to get a floor of
second.