Simplify stratum and accounting modules
They were separate crates with this ideal of making them independent
of p2pool. This was creating unnecessary complexity with traits
etc. Instead of accumulating such debt, with a non-goal, we culled the
separate crates. Now all p2pool modules are in the same lib
crate. Only the binary crates will be separate, like cli, node and
later server.