Remove ckpool. Use new Share Block Model
A massive single commit to pull out ckpool interaction. We don't track and broadcast ckpool workbases and jobs anymore.

Instead we swtiched to a model where we use stratum messages from our stratum server. We track pplns share for local node metrics.

We emit ShareBlock which is now composed of ShareHeader, share chain transactions and bitcoin compact block. The share header only includes bitcoin block, the share block has the compact block.

The above is the model we are now working towards. The ShareHeader is working as above. We are not yet embedding the bitcoin compact block in the ShareBlock. That will come in a separate PR once hydrapool is out.

The goal of this commit was to align Hydrapool with the long term plans of moving off ckpool. The ckpool clean up was long pending and this commit does that.

The pulling out of ckpool meant a deep cleanup and refactor and this commit achives that in a single commit.